MAJOR rewrite
basically nothing was working and the gui was like 1000 lines long so i decided to try and clean it up (it is still messy especially with the time crunch and it might? be super inefficient) most features work now as a demo I just need to complete room management which should be pretty quick. If you need the database credentials please email me!